+++ com.unity.render-pipelines.universal/Editor/Deprecated.cs
@@ -14,13 +14,13 @@ namespace UnityEditor.Rendering.LWRP
 
 namespace UnityEditor.Rendering.Universal
 {
-    [Obsolete("ForwardRendererDataEditor has been deprecated. Use StandardRendererDataEditor instead -> StandardRendererDataEditor (UnityUpgradable)", true)]
+    [Obsolete("ForwardRendererDataEditor has been deprecated. Use UniversalRendererDataEditor instead (UnityUpgradable) -> UniversalRendererDataEditor", true)]
     [MovedFrom("UnityEditor.Rendering.LWRP")]
     public class ForwardRendererDataEditor : ScriptableRendererDataEditor
     {
         public override void OnInspectorGUI()
         {
-            throw new NotSupportedException("ForwardRendererDataEditor has been deprecated. Use StandardRendererDataEditor instead");
+            throw new NotSupportedException("ForwardRendererDataEditor has been deprecated. Use UniversalRendererDataEditor instead");
         }
     }